# packages/contracts/openapi/wedge_openapi.yaml
openapi: 3.0.3
info:
  title: Wedge API
  version: 1.0.0-draft
  description: Wedge public REST API (/api), internal callback API (/internal), and canonical transport schemas. This draft
    includes report shares, checkpoint callback, and evidence packet retrieval.
servers:
- url: http://localhost:8080
  description: local API server for Swagger MVP testing
- url: https://api.wedge.example.com
  description: production placeholder
- url: https://staging-api.wedge.example.com
  description: staging placeholder
security:
- HumanBearer: []
...
paths:
...
  /api/runs:
    get:
      tags:
      - Runs
      operationId: listRuns
      summary: 실행 목록 조회
      description: 생성된 Run 목록과 상태 필터링을 확인한다. signup/login 후 Bearer 토큰으로 호출해 생성된 Run이 목록에 보이는지 테스트한다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - name: projectId
        in: query
        schema:
          type: string
          format: uuid
      - name: status
        in: query
        schema:
          type: string
      - name: analysisStatus
        in: query
        schema:
          type: string
      - name: createdFrom
        in: query
        schema:
          type: string
          format: date-time
      - name: createdTo
        in: query
        schema:
          type: string
          format: date-time
      responses:
        '200':
          description: Run list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRunResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
    post:
      tags:
      - Runs
      operationId: createRun
      summary: 실행 생성
      description: 분석 실행 단위를 생성한다. scenarioPlan을 생략하면 기본 사용자 실행으로 agent.execute.request를 발행하고, scenarioPlan을 포함하면 scripted/replay 실행으로 run.execute.request를 발행한다. Swagger의 replay 예시는 projectId와 scenarioTemplateVersionId를 로컬 DB/seed 값으로 교체해야 실제 테스트가 가능하다.
      security:
      - HumanBearer: []
      parameters:
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunCreateRequest'
            examples:
              agentDefault:
                summary: Runner Agent 기본 실행 생성
                description: scenarioPlan 없이 생성하면 Run 시작 시 agent.execute.request 공식 경로로 발행된다.
                value:
                  projectId: 8f06dca8-9c4d-4f20-b1a8-1d5ee40a9923
                  name: Checkout entry Agent 실행
                  startUrl: https://example.com/product/sample
                  goal: 결제나 최종 주문 제출 없이 checkout 진입 경로를 확인한다.
                  devicePreset: desktop
                  scenarioOverrides:
                    source: swagger-agent-default
              signupBasicMvp:
                summary: Runner smoke 실행 생성
                description: infra/scripts/seed-real-run-smoke.mjs가 생성하는 projectId/scenarioTemplateVersionId를 사용하는 최소 Runner
                  연동 테스트 예시다. target URL을 바꾸면 startUrl과 scenarioPlan.start_url을 같은 값으로 맞춘다.
                value:
                  projectId: 8f06dca8-9c4d-4f20-b1a8-1d5ee40a9923
                  name: Runner smoke 실행
                  startUrl: https://example.com/
                  goal: Runner가 URL을 열고 checkpoint/artifact/evidence callback을 API 서버로 보내는지 확인한다.
                  devicePreset: desktop
                  scenarioTemplateVersionId: 5c5f4c77-0c32-4ab3-9841-2b6f6cc07a40
                  scenarioOverrides:
                    source: swagger-runner-smoke
                  scenarioPlan:
                    schema_version: '0.5'
                    plan_id: smoke_landing_cta_hero_only
                    scenario_type: custom_compiled
                    template_key: landing-cta
                    goal: 'Runner smoke: URL open, checkpoint, evidence callback, API evidence packet.'
                    start_url: https://example.com/
                    environment:
                      device: desktop
                      viewport:
                        width: 1440
                        height: 900
                      locale: ko-KR
                      timezone: Asia/Seoul
                      permissions: []
                      auth_state: anonymous
                      depth_id: hero-only
                    safety:
                      allow_external_navigation: false
                      allow_payment_commit: false
                      allow_destructive_action: false
                      use_synthetic_inputs: true
                      stop_before_real_payment: true
                    steps:
                    - step_id: step_001_goto_start_url
                      stage: FIRST_VIEW
                      description: Smoke 대상 URL 첫 화면 열기
                      action:
                        type: goto
                        target:
                          url: https://example.com/
                      settle_strategy:
                        type: network_idle
                        timeout_ms: 5000
                      checkpoint: true
                    - step_id: step_002_checkpoint_landing_cta
                      stage: CTA
                      description: Landing CTA evidence checkpoint
                      action:
                        type: checkpoint
                        target:
                          scenario_id: landing-cta
                          depth_id: hero-only
                          text: hero section, primary button, nav CTA
                      settle_strategy:
                        type: fixed_short
                        timeout_ms: 500
                      checkpoint: true
      responses:
        '201':
          description: Run created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleRunResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/runs/{runId}:
    get:
      tags:
      - Runs
      operationId: getRun
      summary: 실행 상세 조회
      description: runId로 현재 실행 상태, 시작/종료 시각, 분석 상태를 확인한다. createRun 또는 listRuns에서 받은 id로 테스트한다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Run detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleRunResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags:
      - Runs
      operationId: deleteRun
      summary: 실행 삭제
      description: runId의 실행을 삭제 처리한다. MVP 확인 중 데이터 보존이 필요하면 호출하지 않는다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      responses:
        '204':
          description: Run deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/runs/{runId}/start:
    post:
      tags:
      - Runs
      operationId: startRun
      summary: 실행 시작
      description: CREATED 상태의 Run을 QUEUED로 전환하고 Runner 실행 요청 outbox 발행을 확인한다. createRun 응답의 data.id로 호출한다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      responses:
        '202':
          description: Run start accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/runs/{runId}/agent/start:
    post:
      tags:
      - Runs
      operationId: startAgentRun
      summary: Agent 실행 시작
      description: CREATED 상태의 Run을 QUEUED로 전환하고 AgentTask 기반 Runner Agent 실행 요청 outbox를 발행한다. 이전 성공 AgentTrace가 있으면 replay_hints로 주입한다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      responses:
        '202':
          description: Agent run start accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/runs/{runId}/stop:
    post:
      tags:
      - Runs
      operationId: stopRun
      summary: 실행 중지 요청
      description: 진행 중인 Run에 중지 요청을 보낸다. 실행 상태 전이 검증이나 수동 중단 테스트에 사용한다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunActionRequest'
      responses:
        '202':
          description: Run stop accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /api/runs/{runId}/steps:
    get:
      tags:
      - Steps
      operationId: listRunSteps
      summary: 실행 단계 목록 조회
      description: Run의 시나리오 step 목록과 각 step의 실행 상태/실패 정보를 조회하는 API다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Run steps
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListStepResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/runs/{runId}/steps/{stepId}:
    get:
      tags:
      - Steps
      operationId: getRunStep
      summary: 실행 단계 상세 조회
      description: 특정 stepId의 상태를 조회하는 계약 API다. stepId는 단계 목록 또는 DB 저장값을 기준으로 테스트한다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - $ref: '#/components/parameters/StepId'
      responses:
        '200':
          description: Run step detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleStepResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/runs/{runId}/events:
    get:
      tags:
      - Events
      operationId: listRunEvents
      summary: 실행 이벤트 목록 조회
      description: Runner가 기록한 실행 이벤트 타임라인을 저장된 이벤트 기준으로 조회하는 API다. stepId와 eventType으로 필터링할 수 있고, cursor는 이전 응답 meta.nextCursor의 이벤트 ID(UUID)를 사용한다. limit은 1~100 범위이며 기본값은 20이다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - name: stepId
        in: query
        schema:
          type: string
          format: uuid
      - name: eventType
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Run events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRunEventResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/runs/{runId}/artifacts:
    get:
      tags:
      - Artifacts
      operationId: listRunArtifacts
      summary: 실행 아티팩트 목록 조회
      description: 스크린샷/HTML/trace 같은 저장 아티팩트 목록과 content URL을 확인한다. Runner callback 이후 결과 확인에 사용한다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - name: type
        in: query
        schema:
          type: string
      - name: stepId
        in: query
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Run artifacts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListArtifactResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/runs/{runId}/artifacts/{artifactId}/content:
    get:
      tags:
      - Artifacts
      operationId: getRunArtifactContent
      summary: 실행 아티팩트 파일 다운로드
      description: artifactId의 실제 파일 바이너리를 내려받는 API다. listRunArtifacts 응답의 artifactId로 이미지/HTML 파일 접근을 확인한다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - $ref: '#/components/parameters/ArtifactId'
      responses:
        '200':
          description: Artifact binary content
          content:
            image/png:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
              description: Legacy PDF report export content is returned as an attachment with Content-Disposition.
            text/markdown:
              schema:
                type: string
              description: Markdown report export content is returned as a UTF-8 attachment with Content-Disposition.
            text/html:
              schema:
                type: string
            application/json:
              schema:
                type: object
                additionalProperties: true
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/runs/{runId}/signals:
    get:
      tags:
      - Signals
      operationId: listRunSignals
      summary: 이슈 시그널 목록 조회
      description: 분석 결과에서 사용자에게 보여줄 이슈 신호를 조회하는 계약 API다. 현재 MVP에서는 report API가 우선 검증 대상이다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - name: stepId
        in: query
        schema:
          type: string
          format: uuid
      - name: category
        in: query
        schema:
          type: string
      - name: source
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Run issue signals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSignalResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /internal/runner/runs/{runId}/artifacts:
    post:
      tags:
      - InternalRunner
      operationId: runnerArtifacts
      summary: Runner 아티팩트 callback
      description: Runner가 캡처한 파일/스크린샷 메타데이터를 저장하는 내부 callback이다. 이후 artifacts API에서 조회되는지 테스트한다.
      security:
      - InternalBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - name: X-Worker-Id
        in: header
        required: true
        schema:
          type: string
      - name: X-Event-Id
        in: header
        required: true
        schema:
          type: string
      - name: X-Signature
        in: header
        description: HMAC-SHA256 digest of the raw JSON request body using the shared runner callback signature secret. The preferred format is `hmac-sha256=<hex>`; raw hex is accepted for compatibility.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunnerArtifactsRequest'
      responses:
        '200':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /api/runs/{runId}/evidence-packet:
    get:
      tags:
      - Runs
      operationId: getRunEvidencePacket
      summary: 실행 EvidencePacket 조회
      description: Run에서 수집된 checkpoint/observation/artifact 기반 EvidencePacket을 조회한다. 분석 요청 전후 evidence shape 확인에 사용한다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: EvidencePacket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleEvidencePacketResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.wedge.example.com/oauth2/authorize
          tokenUrl: https://auth.wedge.example.com/oauth2/token
          scopes:
            wedge.read: Read Wedge resources
            wedge.execute: Create and execute runs
            wedge.export: Export reports
            wedge.admin: Administrative actions
        clientCredentials:
          tokenUrl: https://auth.wedge.example.com/oauth2/token
          scopes:
            wedge.read: Read Wedge resources
            wedge.execute: Create and execute runs
            wedge.export: Export reports
            wedge.admin: Administrative actions
    InternalBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    HumanBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    Unauthorized:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    RunCreateRequest:
      type: object
      properties:
        projectId:
          type: string
          format: uuid
        name:
          type: string
        startUrl:
          type: string
          format: uri
        sourceDiscoveryId:
          type: string
          nullable: true
        goal:
          type: string
          nullable: true
        devicePreset:
          type: string
          enum:
          - desktop
          - tablet
          - mobile
        scenarioTemplateVersionId:
          type: string
          format: uuid
          nullable: true
        scenarioOverrides:
          type: object
          additionalProperties: true
        scenarioPlan:
          type: object
          additionalProperties: true
      required:
      - projectId
      - name
      - startUrl
      - devicePreset
    RunActionRequest:
      type: object
      properties:
        reason:
          type: string
          nullable: true
    AckResponse:
      type: object
      properties:
        data:
          type: object
          nullable: true
          additionalProperties: true
        meta:
          $ref: '#/components/schemas/Meta'
      required:
      - data
      - meta
...
    SingleRunResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Run'
        meta:
          $ref: '#/components/schemas/Meta'
      required:
      - data
      - meta
    ListRunResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Run'
        meta:
          $ref: '#/components/schemas/Meta'
      required:
      - data
      - meta
...